refactor(profiling): move periodic to ddtrace.internal (#2184)
