+++ tests/tracer/test_periodic.py
@@ -3,8 +3,8 @@ import threading
 
 import pytest
 
+from ddtrace.internal import periodic
 from ddtrace.internal import service
-from ddtrace.profiling import _periodic
 
 
 if os.getenv("DD_PROFILE_TEST_GEVENT", False):
@@ -50,7 +50,7 @@ def test_periodic():
     def _on_shutdown():
         x["DOWN"] = True
 
-    t = _periodic.PeriodicRealThreadClass()(0.001, _run_periodic, on_shutdown=_on_shutdown)
+    t = periodic.PeriodicRealThreadClass()(0.001, _run_periodic, on_shutdown=_on_shutdown)
     t.start()
     thread_started.wait()
     thread_continue.set()
@@ -68,7 +68,7 @@ def test_periodic_double_start():
     def _run_periodic():
         pass
 
-    t = _periodic.PeriodicRealThreadClass()(0.1, _run_periodic)
+    t = periodic.PeriodicRealThreadClass()(0.1, _run_periodic)
     t.start()
     with pytest.raises(RuntimeError):
         t.start()
@@ -88,7 +88,7 @@ def test_periodic_error():
     def _on_shutdown():
         x["DOWN"] = True
 
-    t = _periodic.PeriodicRealThreadClass()(0.001, _run_periodic, on_shutdown=_on_shutdown)
+    t = periodic.PeriodicRealThreadClass()(0.001, _run_periodic, on_shutdown=_on_shutdown)
     t.start()
     thread_started.wait()
     thread_continue.set()
@@ -99,13 +99,13 @@ def test_periodic_error():
 
 def test_gevent_class():
     if os.getenv("DD_PROFILE_TEST_GEVENT", False):
-        assert isinstance(_periodic.PeriodicRealThreadClass()(1, sum), _periodic._GeventPeriodicThread)
+        assert isinstance(periodic.PeriodicRealThreadClass()(1, sum), periodic._GeventPeriodicThread)
     else:
-        assert isinstance(_periodic.PeriodicRealThreadClass()(1, sum), _periodic.PeriodicThread)
+        assert isinstance(periodic.PeriodicRealThreadClass()(1, sum), periodic.PeriodicThread)
 
 
 def test_periodic_service_start_stop():
-    t = _periodic.PeriodicService(1)
+    t = periodic.PeriodicService(1)
     t.start()
     with pytest.raises(service.ServiceAlreadyRunning):
         t.start()
@@ -118,11 +118,11 @@ def test_periodic_service_start_stop():
 
 
 def test_periodic_join_stop_no_start():
-    t = _periodic.PeriodicService(1)
+    t = periodic.PeriodicService(1)
     t.join()
     t.stop()
     t.join()
-    t = _periodic.PeriodicService(1)
+    t = periodic.PeriodicService(1)
     t.stop()
     t.join()
     t.stop()
@@ -132,5 +132,5 @@ def test_is_alive_before_start():
     def x():
         pass
 
-    t = _periodic.PeriodicRealThreadClass()(1, x)
+    t = periodic.PeriodicRealThreadClass()(1, x)
     assert not t.is_alive()
